Dashboard now fully implements MVP
